+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Plugin developers:
 - add plugin config validation functionality
 - switch from a top menu to a sidebar menu
 - show pairing status for bridges around the ui
+- refactor status and child bridge widgets
 
 ### Other Changes